Move all the component classes from the Widgets/Variance directories to a new Components directory.
