+++ Components/SegmentRow.py
