+++ Variance/varianceApp.py
@@ -2,7 +2,7 @@
 
 import tkinter as tk
 import threading
-from Variance import VarianceColumn
+from Components import VarianceColumn
 from DataClasses import AllSplitNames
 from Variance import varianceCalculator as varcalc
